fix(view-tests): capture reconcile events through one shared, race-free fixture
global_key_reparent's active_to_active test intermittently captured ZERO
events under full-workspace test load (2 failures in 3 local `just ci`
runs). Cause: tracing caches per-callsite interest globally; while no
interested subscriber exists a callsite is cached `never`, and every
scoped `with_default` dispatcher registration/drop on any other test
thread rebuilds that cache — an event emitted in the `never` window is
skipped before the emitting thread's own dispatcher is consulted.

production_reconcile_emits already carried the antidote (a process-global
always-enabled no-op Registry, keeping every callsite's interest
non-`never`) — but only for itself, and only if one of its tests happened
to run first. The fixture now lives in a shared module both files use, so
the guarantee no longer depends on test order.

Validated: view_it 268/268 across 12 consecutive runs plus a full
`just ci`; the mechanism (no `never`-interest window with a permanent
global subscriber) is order-independent.
